requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-chroma>=0.2.0
langchain-community>=0.0.10
chromadb>=0.4.0
openai>=1.0.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
